+++ SnakeController/Cell.hpp
@@ -1,0 +1,6 @@
+enum Cell
+{
+    Cell_FREE,
+    Cell_FOOD,
+    Cell_SNAKE
+};
